Added incomplete & broken efficient two-step impl.
Broken because I couldn't figure out how exactly it worked - I think
that the slow two-step implementation may also be broken.

Additionally, such an implementation is not necessary for our project,
as we will be using the Loeffler implementation for maximum efficiency
in both our C code and in our optimized assembly/custom-firmware
implementations.