Implemented function which actually runs dct
main() runs dct(), which takes the image from main, and passes it
through run_dct() four times. run_dct() takes the image data and a
function pointer, and runs the passed function on each 8x8 block of the
image, printing out the results. The four functions that are executed on
each 8x8 block of the image are:

- A function which copies the input directly to the output. This is used
  only so that the input is printed for comparison purposes.

- A naive implementation of dct, using quadruply-nested for loops and a
  cos(x*pi/16) lookup table. This is used as reference, as it is likely
  the most bug-free version of the three actual implementations (though
  there is no guarantee that it is actually working correctly).

- A simple version of dct using matrix multiplications. This is
  somewhat more efficient than the truly naive version, but is more
  prone to errors. I wasn't able to get the efficient
  matrix-multiplication version working, but it shouldn't be too hard,
  though there isn't much point to it.

- A more complicated version of dct using the Loeffler method. This is
  the most efficient 1-D dct implementation we know about, so we're
  using it. However, this currently seems to be broken, even though as
  far as I can tell, all of the operations are performed correctly as
  according to the diagram and details presented on the slides. The
  issues may relate to either the fact that floating-point operations
  are used, or that the algorithm is incorrectly implemented at a higher
  level than the 1-D implementation (i.e. it may be that the 1-D
  implementation is working correctly but is being called incorrectly).

Additionally, all of these implementations are fundamentally unoptimized
and (from the perspective of this course) unusable, because they use
double-precision floating-point arithmetic for all of their operations.
However, the interfaces to these functions use signed 8-bit integers, so
we would only need to fix the internals in order to fix them. Frankly, I
didn't pay enough attention in the class where we covered how to
maximize precision with trigonometric functions, so I'm not sure how to
do it (though we don't actually need trigonometric functions - only
constant values that trigonometric outputs - so we may be able to get
away with not learning that stuff). This is a critical problem that will
need to be addressed before starting to generate and optimize assembly.

Another issue is whether the inputs & outputs should be signed or
unsigned. Online info seems to be unclear. It would seem that unsigned
inputs could result in negative outputs, which may lead to problems.
Perhaps internally we could use signed values, then output unsigned
values if we know that the results will be unsigned - but I'm not
confident that that would be the case.

A final improvement (which may fix the issue with the Loeffler method,
but that's not clear) will be to create a Butterfly() routine, as
directed by the assignment slides, rather than running it directly
within the Loeffler 1-D dct function. This will also allow us to more
clearly optimize that routine out into an assembly instruction.